Addresses issue #32 and made updates to form dialogs to now use bootstrap 3 panel styles
Added a form to request that changes are saved when a user is navigating away from a project that contains unsaved changes
Fixed issue where modifying the pvsfiles or the widgets did not mark the project as dirty. Consequently, WidgetManager now fires a WidgetModified event.
Also fixes issue #44 where callback function in projectManager’s saveProject is not invoked