Fix bug in file tree where folders immediate collapsed after opening
This fix adds a separate "ItemExpanded" event that loads the folder's
children, rather than it being done when the item is selected for the
first time. This also makes the behaviour more consistent - previously
the first single-click on a folder would expand it, but subsequent
clicks would not. Now clicking the expander arrow or double clicking
expands/collapses, while single clicks on the item only ever select it.

Also combined some related repeated code into a function.