Refactor CSS to SCSS.
The main aim of this is to give the project's styling more structure and
organisation, making it easier to maintain, modify and add to the
styling.

The SCSS compilation is handled by node-sass and autoprefixer via a
Grunt task. A section has been added to the project's readme
describing how to use this task. Currently the compiled css file
(style.css) is committed to the Git repo. This means it is not necessary
to compile the css when simply running PVSio-Web (i.e. ./start.sh), but
if this causes any issues the start script could easily be modified to
also run `grunt sass && grunt postcss` as part of the installation.

A readme file within the css folder describes the SCSS directory
structure, but the approach is essentially the standard approach that is
often used with SASS/LESS/etc.

This commit only makes minor changes to the actual content of the CSS -
for the most part it is all identical to the original CSS.