Refactor styling - introduce variables and restructure some styles
This doesn't alter the appearance of the actual UI (excluding a few
small cases where things such as colours have been tweaked extremely
slightly to be more consistent), but extracts common values such as
colours and font sizes out to variables and common styles to
mixins/placeholders. This should help ensure consistency and make it
easier to tweak common styles.

A couple of components have been reworked to follow better styling
practices, e.g. so that they no longer target several IDs (which makes
it hard to add new elements that share the style and to override the
styling with modifier classes) and instead the elements share a common
class that is styled instead. These changes were made to the Emulink
tabs and prototype builder tabs.