Start on modularizing the WidgetsListView into a Backbone view
Most of the functionality of the old view is preserved, although a few
things still need to be converted/implemented - these are mainly reliant
on the rest of the prototype builder being converted to Backbone/a more
modular structure.

The purpose of this is to make ite easier/possible to implement
different prototyping systems using the same base components.