Refactor more of the existing prototype builder
Moved the imagemapper initialisation to the PrototypeImageView (rather
than PrototypeBuilder) and made some changes so that it is not reliant
on any particular WidgetManager or PrototypeBuilder.

WidgetManager has been refactored to extend an agnostic base class (i.e.
the base class provides core functionality that any type of
WidgetManager would need). WidgetManager now also uses Backbone.Events,
as it already provides functionality for effortlessly removing event
listeners (avoiding memory leaks with the views) and there was little
apparent reason not to. The other refactored classes have also been
changed to not rely on the singleton WidgetManager and instead take it
as an argument (allowing other WidgetManagers to be passed in).

Fixed several bugs created by the earlier refactoring (or bugs that
became apparent via the changes) as well. Only know bug that has been
introduced is that images in the prototype builder do not resize.