Modify question dialog so that the button importance/level can be set
This allows things such as delete confirmation to colour the "delete"
button to reflect its severity/importance. I've also changed the
template so that the cancel button is not emphasised, giving more focus
to the default action (see http://ux.stackexchange.com/q/49991 for
discussion related to this sort of button colouring).