Fix delete key only being picked up by the last opened prototyper
Both imagemapper instances were registering a keydown listener on the
page body, so only the last panel to be added actually received the
event. This has been changed so that the main PVSioWeb class listens for
key events and passes them onto the 'active' panel (if the panel
provides the appropriate handler function). The active panel is
whichever was last moused-over.