Implement saving PIM prototypes.
This builds on the existing project management functionality, so the PIM
prototype is saved whenever the overall project is saved.
Modified the Descriptor class to support generating its content lazily -
i.e. only when it is requested. This saves having to serialize the PIM
prototype every time it is modified and instead only serialize it when
actually saving.