Add simple function for exporting PM as text from the chart editor
This allows the model to be exported as a textual representation of a
presentation model, containing several component models (for each
state). The format follows the same structure as the standard
representation of PMs used elsewhere (i.e. in papers).
The export code is loosely based on the existing test generation code.